add another unit test and some refactoring
